requirements.txt.txt
python-dotenv
deepsearch-toolkit
tesseract
pytesseract
pdf2image
pillow